Refactor/reorganize into a new class dynamic.
Players have a parent Player class and two classes that inherit from it: the UserPlayer and the ComputerPlayer. Whether the player is the program or the user is the central dynamic NOT what role the player is currently filling. This will make the logic of the player actions easier to use and more self-contained